--- Main.java
@@ -1,9 +1,0 @@
-package testjava;
-import java.util.*;
-
-public class Main {
-    public static void main(String[] args) {
-        List<Integer> sequencia = Fibonacci.gerarSequencia(9);
-        System.out.println(sequencia);
-    }
-}
