+++ src/exercicio7/Main.java
@@ -1,0 +1,9 @@
+package exercicio7;
+import java.util.*;
+
+public class Main {
+    public static void main(String[] args) {
+        List<Integer> sequencia = Fibonacci.gerarSequencia(9);
+        System.out.println(sequencia);
+    }
+}
